// src/app/shared/models/permission.interface.ts
export interface PermissionInterface {
  dashboard: boolean;
  eCommerce: boolean;
  liveTraffic: boolean;
  ngBootstrap: boolean;
  layoutBuilder: boolean;
  userManage: boolean;
  companyManage: boolean;
  tagManage: boolean;
  reportManage: boolean;
  publisherReportingManage: boolean;
  apiDocumentationManage: boolean;
  protectedMedia: boolean;
  notifications: boolean;
  role: number;
  _id: string;
  _key: string;
  _rev: string;
}